requirements.txt
-i https://pypi.org/simple
annotated-doc==0.0.4; python_version >= '3.8'
anyio==4.11.0; python_version >= '3.9'
bcrypt==5.0.0; python_version >= '3.8'
cachetools==6.2.2; python_version >= '3.9'
certifi==2025.11.12; python_version >= '3.7'
cffi==2.0.0; python_version >= '3.9'
charset-normalizer==3.4.4; python_version >= '3.7'
click==8.1.8; python_version >= '3.7'
cryptography==46.0.3; python_version >= '3.8' and python_full_version not in '3.9.0, 3.9.1'
ecdsa==0.19.1; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
exceptiongroup==1.3.1; python_version >= '3.7'
fastapi==0.122.0; python_version >= '3.8'
google-ai-generativelanguage==0.6.15; python_version >= '3.7'
google-api-core[grpc]==2.28.1; python_version >= '3.7'
google-api-python-client==2.187.0; python_version >= '3.7'
google-auth==2.43.0; python_version >= '3.7'
google-auth-httplib2==0.2.1; python_version >= '3.7'
google-generativeai==0.8.5; python_version >= '3.9'
googleapis-common-protos==1.72.0; python_version >= '3.7'
greenlet==3.2.4; python_version >= '3.9'
grpcio==1.76.0; python_version >= '3.9'
grpcio-status==1.71.2; python_version >= '3.9'
h11==0.16.0; python_version >= '3.8'
httplib2==0.31.0; python_version >= '3.6'
idna==3.11; python_version >= '3.8'
passlib[bcrypt]==1.7.4
proto-plus==1.26.1; python_version >= '3.7'
protobuf==5.29.5; python_version >= '3.8'
psycopg2-binary==2.9.11; python_version >= '3.9'
pyasn1==0.6.1; python_version >= '3.8'
pyasn1-modules==0.4.2; python_version >= '3.8'
pycparser==2.23; python_version >= '3.8'
pydantic==1.10.24; python_version >= '3.7'
pyparsing==3.2.5; python_version >= '3.9'
python-dotenv==1.2.1; python_version >= '3.9'
python-jose[cryptography]==3.5.0; python_version >= '3.9'
python-multipart==0.0.20; python_version >= '3.8'
requests==2.32.5; python_version >= '3.9'
rsa==4.9.1; python_version >= '3.6' and python_version < '4'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.44; python_version >= '3.7'
starlette==0.49.3; python_version >= '3.9'
tqdm==4.67.1; python_version >= '3.7'
typing-extensions==4.15.0; python_version >= '3.9'
uritemplate==4.2.0; python_version >= '3.9'
urllib3==2.5.0; python_version >= '3.9'
uvicorn==0.38.0; python_version >= '3.9'
